Refactor app with seperate component for the Header component
