+++ src/app.js
@@ -2,6 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import AddOption from './components/AddOption';
 import Option from './components/Option';
+import Header from './components/Header';
 
 class IlearnxApp extends React.Component {
     constructor(props){
@@ -93,20 +94,6 @@ class IlearnxApp extends React.Component {
             </div>
         );
     }
-}
-
-// Header stateless functional component for User
-const Header = (props) => {
-    return (
-        <div>
-            <h1>{props.title}</h1>
-            {props.subtitle && <h2>{props.subtitle}</h2>}
-        </div>
-    );
-}
-
-Header.defaultProps = {
-    title: 'iLearnx App'
 }
 
 // Action stateless functional component for User
