+++ src/components/Header.js
@@ -1,0 +1,17 @@
+import React from 'react';
+
+// Header stateless functional component for User
+const Header = (props) => {
+    return (
+        <div>
+            <h1>{props.title}</h1>
+            {props.subtitle && <h2>{props.subtitle}</h2>}
+        </div>
+    );
+}
+
+Header.defaultProps = {
+    title: 'iLearnx App'
+}
+
+export default Header;
